requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
bcrypt==4.1.3
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
comm==0.2.2
cryptography==42.0.7
debugpy==1.8.1
decorator==5.1.1
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
executing==2.0.1
fastapi==0.110.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
ipykernel==6.29.4
ipython==8.24.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
orjson==3.10.1
packaging==24.0
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.1
pydantic-extra-types==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
pyzmq==26.0.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
stack-data==0.6.3
starlette==0.37.2
tornado==6.4
traitlets==5.14.3
typing_extensions==4.11.0
ujson==5.9.0
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
